discord_against_humanity/bot/requirements.txt
discord.py
motor
html2text
async-property